Rename package printer to output
