+++ pkg/output/escape.go
@@ -1,4 +1,4 @@
-package printer
+package output
 
 import "strings"
