+++ pkg/output/printer_test.go
@@ -1,0 +1,7 @@
+package output
+
+import "testing"
+
+func Test_interface(t *testing.T) {
+	var _ Interface = &Printer{}
+}
